+++ packages/orq-rc/RELEASES.md
@@ -1509,3 +1509,13 @@ Based on:
 - [typescript v3.6.0-rc.7] packages/orq-rc
 ### Releases
 - [NPM v3.6.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/3.6.0-rc.7 - packages/orq-rc
+
+## 2025-05-16 08:03:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.546.0 (2.603.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.6.0-rc.8] packages/orq-rc
+### Releases
+- [NPM v3.6.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/3.6.0-rc.8 - packages/orq-rc